main: assume PA is equal to VA
Since the job of a TLB is to deliver a physical frame number
corresponding to a virtual page number, assume that the physical address
is equal to the virtual address irrespective of the instruction type,
i.e., read or write.

This assumption is made for the simplicity of the simulation and is in
fact technically inaccurate.

Co-authored-by: Abhay Kumar <abhay98394@gmail.com>
Co-authored-by: Shourya Shukla <periperidip@gmail.com>